requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
motor>=3.3.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pytest>=7.4.0
httpx>=0.24.1

# Storage
boto3>=1.28.0

# AI & Parsing Dependencies
pdfplumber>=0.10.0
PyPDF2>=3.0.0
spacy>=3.7.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
